% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{cf.csur}
\alias{cf.csur}
\title{Compute reduction in contour-distance}
\usage{
cf.csur(objMean, objSd, nugget)
}
\arguments{
\item{nugget}{the noise variance to compute the ALC factor}

\item{objMean:}{predicted mean response}

\item{objSd:}{posterior standard deviation of the response}
}
\description{
cSUR for Contour Finding
}
\details{
compute the change in ZC = sd*(1-sqrt{(nugget)})/sqrt{(nugget + sd^2)}
}
