% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{cf.el}
\alias{cf.el}
\title{Compute expected loss using the optimal stopping loss function.}
\usage{
cf.el(objMean, objSd)
}
\arguments{
\item{objMean:}{predicted mean response}

\item{objSd:}{posterior standard deviation of the response}
}
\description{
Expected Loss for Contour Finding
}
