% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{maxi.call.payoff}
\alias{maxi.call.payoff}
\title{Max Call payoff}
\usage{
maxi.call.payoff(x, model)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{x:}{asset prices}
}
\description{
Multivariate max call
}
