% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{mini.put.payoff}
\alias{mini.put.payoff}
\title{Min Put payoff}
\usage{
mini.put.payoff(x, model)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{x:}{asset prices}
}
\description{
Multivariate min Put
}
