% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{treeDivide.BW}
\alias{treeDivide.BW}
\title{Create a Bouchard-Warin equal prob grid}
\usage{
treeDivide.BW(grid, curDim, model, test)
}
\arguments{
\item{grid}{-- dataset of x-values}

\item{curDim}{dimension of the grid}

\item{test}{testPaths to predict along as well}
}
\description{
Recursively sort along each of the d-coordinates
At the end do local linear regression at each leaf
This is a recursive algorithm!
first column is reserved for the y-coordinate (timingValue)
It's safest if nrows(grid) is divisible by model$nChildren^dim
}
